fix minor bug and add favourite screen

+++ app/src/main/java/com/example/storyefun/data/models/Chapter.kt
@@ -3,6 +3,6 @@ package com.example.storyefun.data.models
 data class Chapter(
     val id: String = "",        // Default value
     val title: String = "",      // Default value
-    val order: Int = 0,
+    val order: Long = 0,
     val content: List<String> = emptyList()
 )
